Delete Controller の疎結合
